requirements.txt
# Web scraping and automation
selenium>=4.15.2
webdriver-manager>=4.0.1
requests>=2.31.0

# Image processing
Pillow>=10.1.0
opencv-python>=4.8.1.78
python-doctr>=0.7.0
python-doctr[torch,viz,html,contrib]

# Translation
deep-translator>=1.11.4

# Data processing
numpy>=1.24.0
scikit-learn>=1.3.2

# Development
pytest>=7.4.3
black>=23.11.0